move multipart to entity package
