+++ library/entity/multipart/src/main/java/org/quiltmc/qsl/entity/multipart/api/AbstractEntityPart.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.multipart.api;
+package org.quiltmc.qsl.entity.multipart.api;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityDimensions;
