+++ library/entity/multipart/src/main/java/org/quiltmc/qsl/entity/multipart/mixin/EnderDragonPartMixin.java
@@ -14,20 +14,25 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.multipart.api;
+package org.quiltmc.qsl.entity.multipart.mixin;
 
-import net.minecraft.entity.Entity;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+import net.minecraft.entity.boss.dragon.EnderDragonEntity;
 import net.minecraft.entity.boss.dragon.EnderDragonPart;
 
-import org.quiltmc.qsl.base.api.util.InjectedInterface;
+import org.quiltmc.qsl.entity.multipart.api.EntityPart;
 
-/**
- * Represents the sub-parts of a {@link MultipartEntity}.
- *
- * @param <E> the {@link Entity} that owns this {@link EntityPart}
- * @see EnderDragonPart
- */
-@InjectedInterface(EnderDragonPart.class)
-public interface EntityPart<E extends Entity> {
-	E getOwner();
+@Mixin(EnderDragonPart.class)
+public class EnderDragonPartMixin implements EntityPart<EnderDragonEntity> {
+	@Shadow
+	@Final
+	public EnderDragonEntity owner;
+
+	@Override
+	public EnderDragonEntity getOwner() {
+		return this.owner;
+	}
 }
