+++ library/entity/multipart/src/main/java/org/quiltmc/qsl/entity/multipart/mixin/PlayerEntityMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.multipart.mixin;
+package org.quiltmc.qsl.entity.multipart.mixin;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -23,7 +23,7 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
 
-import org.quiltmc.qsl.multipart.api.EntityPart;
+import org.quiltmc.qsl.entity.multipart.api.EntityPart;
 
 @Mixin(PlayerEntity.class)
 public class PlayerEntityMixin {
