+++ library/entity/multipart/src/main/java/org/quiltmc/qsl/entity/multipart/mixin/client/ClientEntityHandlerMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.multipart.mixin.client;
+package org.quiltmc.qsl.entity.multipart.mixin.client;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -28,9 +28,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.Entity;
 
-import org.quiltmc.qsl.multipart.api.EntityPart;
-import org.quiltmc.qsl.multipart.api.MultipartEntity;
-import org.quiltmc.qsl.multipart.impl.EntityPartTracker;
+import org.quiltmc.qsl.entity.multipart.api.EntityPart;
+import org.quiltmc.qsl.entity.multipart.api.MultipartEntity;
+import org.quiltmc.qsl.entity.multipart.impl.EntityPartTracker;
 
 @Environment(EnvType.CLIENT)
 @Mixin(targets = "net/minecraft/client/world/ClientWorld$ClientEntityHandler")
@@ -43,7 +43,7 @@ public class ClientEntityHandlerMixin {
 	private void startTrackingEntityParts(Entity entity, CallbackInfo ci) {
 		if (entity instanceof MultipartEntity multipartEntity) {
 			for (EntityPart<?> part : multipartEntity.getEntityParts()) {
-				((EntityPartTracker) this.field_27735).getEntityParts().put(((Entity) part).getId(), (Entity) part);
+				((EntityPartTracker) this.field_27735).quilt$getEntityParts().put(((Entity) part).getId(), (Entity) part);
 			}
 		}
 	}
@@ -52,7 +52,7 @@ public class ClientEntityHandlerMixin {
 	private void stopTrackingEntityParts(Entity entity, CallbackInfo ci) {
 		if (entity instanceof MultipartEntity multipartEntity) {
 			for (EntityPart<?> part : multipartEntity.getEntityParts()) {
-				((EntityPartTracker) this.field_27735).getEntityParts().remove(((Entity) part).getId(), part);
+				((EntityPartTracker) this.field_27735).quilt$getEntityParts().remove(((Entity) part).getId(), part);
 			}
 		}
 	}
