+++ library/entity/multipart/src/main/resources/assets/quilt_entity_multipart/icon.png
