+++ library/entity/multipart/src/main/resources/quilt_entity_multipart.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "org.quiltmc.qsl.multipart.mixin",
+  "package": "org.quiltmc.qsl.entity.multipart.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "EnderDragonEntityMixin",
