+++ library/entity/multipart/src/testmod/resources/quilt.mod.json
@@ -1,18 +1,19 @@
 {
   "schema_version": 1,
   "quilt_loader": {
-    "group": "org.quiltmc.qsl.multipart",
-    "id": "qsl_multipart_testmod",
+    "group": "org.quiltmc.qsl.entity.multipart",
+    "id": "qsl_entity_multipart_testmod",
     "version": "1.0.0",
     "metadata": {
-      "name": "Quilt Multipart API Test Mod",
+      "name": "Quilt Entity Multipart API Test Mod",
       "license": "Apache-2.0"
     },
     "intermediate_mappings": "net.fabricmc:intermediary",
     "load_type": "always",
     "depends": [
-      "quilt_loader"
+      "quilt_loader",
+      "quilt_entity_multipart"
     ]
   },
-  "mixin": "quilt_multipart_testmod.mixins.json"
+  "mixin": "quilt_entity_multipart_testmod.mixins.json"
 }
