+++ library/entity/multipart/src/testmod/resources/quilt_entity_multipart_testmod.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "org.quiltmc.qsl.multipart.test.mixin",
+  "package": "org.quiltmc.qsl.entity.multipart.test.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "CreeperEntityMixin"
